+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [#90](https://github.com/dblock/jenkins-ansicolor-plugin/pull/90): Added missing handling for ATTRIBUTE_CONCEAL_OFF - [@Joe-Merten](https://github.com/Joe-Merten).
 * [#90](https://github.com/dblock/jenkins-ansicolor-plugin/pull/90): Added support for italic, strikeout, framed and overline attributes - [@Joe-Merten](https://github.com/Joe-Merten).
 * [#92](https://github.com/dblock/jenkins-ansicolor-plugin/pull/92): Fixed high intensity colors for both foreground and background - [@Joe-Merten](https://github.com/Joe-Merten).
+* Added support for negative (inverse colors) attribute - [@Joe-Merten](https://github.com/Joe-Merten).
 * Your contribution here.
 
 0.4.3 (11/20/2016)